IR Outliner Pass
This is the work of River Riddle
https://github.com/River707/llvm.git, outliner branch

With some cleanup and fixes.

Fix IRO calculation of spills count.

Only remove instructions for the interval when their end of use is known.
Calculate the registers used by the end of use / dead interval.
To avoid double counting, if the dead spill is less than the normal spill,
use the dead spill.  And check for any trailing spillage.